% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-helpers-to-build-reference-data.R
\name{add_row_bounding_box}
\alias{add_row_bounding_box}
\title{Title}
\usage{
add_row_bounding_box(data = nc)
}
\arguments{
\item{data}{}
}
\description{
Title
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
add_row_bounding_box(nc)
}
