% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-helpers-to-build-reference-data.R
\name{create_geometries_reference}
\alias{create_geometries_reference}
\title{Title}
\usage{
create_geometries_reference(
  sfdata = nc,
  id_cols = c(NAME, FIPS),
  geometry = geometry
)
}
\arguments{
\item{geometry}{}
}
\description{
Title
}
